Cleanup visitors doc for rewrite8
I didn't really see many changes that needed to be made here and
I'm unsure if that's because none of this stuff changed or if we
are missing some key pieces.

I did find a few minor cleanup pieces and I did call out that one
of the examples doesn't work. I think I'll need some help on that
front.