Add Micronaut 3 -> 4 guide
I'm not familiar with Micronaut at all but I did my best to find
some examples to add to this doc. Many of the examples are taken
from the tests for the recipes that make up this larger recipe.

I chose to include the main ones that the Micronaut Guides needed to
update in their migration.

Fixes: https://github.com/openrewrite/rewrite-docs/issues/217