Fix broken links
Takes the script provided in: https://github.com/openrewrite/rewrite-docs/issues/308

and goes through the docs -- fixing many of the links. Some weren't applicable - and others
need recipe changes or generation changes. I notice that many of the links were because the
recipe is now called "Recipes" at the end so perhaps something else needs to be fixed to
make sure the generation detects that.