Remove spell check plugin
While it has caught a few spelling mistakes, it's also been causing problems as so much of this documentation is auto-generated from stuff that we can't directly control. Furthermore, it's not always obvious to other people coming in why builds fail or how to fix them. Removing it for now - will potentially re-evaluate in the future.